Drop Postgres for DuckDB; reach dbt through one versioned client
- Warehouses are bigquery (production) and duckdb (local work, tests, default compose):
  src/dialects/duckdb.js replaces postgres.js (guarded JSON reads, TRY_CAST, lists, exact
  mergeable sketches, CTE lowering; match_recognize's CTE form takes the dialect).
- src/dbt/: createDbt picks the client by the dbt major version (1.x implemented, v2 refused
  with the reason); one contract (parse/run/seed/show/relationColumns/query/validate/warehouse).
  A DuckDB file admits one process: every dbt/mf process on it takes the warehouse's turn, keyed
  by the database file read from the profile, and the MetricFlow sidecar releases it per request.
- Tests: fixtures on DuckDB (DUCKDB_PATH per test file), a warehouse harness in place of PGlite,
  one venv (dbt-core 1.11 + dbt-duckdb + dbt-metricflow + pandas) for everything incl. the
  python stage; PGlite removed from devDependencies.
- requirements, compose, .env.example, docs and CLAUDE.md follow.

+++ test/integration/fixtures/scd_project/models/fct_events.sql
@@ -5,6 +5,6 @@ select
     event_name,
     device_time,
     event_data,
-    (event_data->>'price_in_usd')::numeric as price_in_usd_of_event_data,
+    (event_data->>'price_in_usd')::double as price_in_usd_of_event_data,
     (event_data->>'product_id')            as product_id_of_event_data
 from {{ ref('seed_events_scd') }}
